% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflam.divide_domain.R
\name{tflamingo.divide_domain}
\alias{tflamingo.divide_domain}
\title{Divide domain for flamingo}
\usage{
tflamingo.divide_domain(
  tflamingo_high_res_obj,
  domain_res,
  frag_res,
  temp_folder
)
}
\arguments{
\item{tflamingo_high_res_obj}{high resolution flamingo object to generate domain}

\item{domain_res}{Size of the domains in bps, e.g. 1e6. Try strawr::readHicBpResolutions() to see available resolutions.}

\item{frag_res}{Size of the fragment in bps, e.g. 1e4. Try strawr::readHicBpResolutions() to see available resolutions.}

\item{temp_folder}{store the temporary domain data}
}
\description{
Divide domain for flamingo
}
