% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflamingo.process_one_file.R
\name{tflamingo.process_one_file}
\alias{tflamingo.process_one_file}
\title{tflamingo.process_one_file}
\usage{
tflamingo.process_one_file(
  input_path,
  resolution,
  chr_name,
  opt_path,
  assembly,
  suffix
)
}
\arguments{
\item{input_path}{Path containing the scHi-C IF matrices of all chromosomes; will search for the IF matrices in the sub-folder named by the chr_name}

\item{resolution}{Resolution of the IF matrices, e.g. 10000.}

\item{chr_name}{chromosome name e.g. "chr19"}

\item{opt_path}{path to write the imputed IF matrices.}

\item{assembly}{Assembly of the scHi-C experiment, e.g. hg38, hg19 and mm10.}
}
\value{
None
}
\description{
Generate the contact matrix from scHi-C contact list
}
\keyword{tFLAMINGO}
